add typecheck support for type parameters.
i can't believe that works. when typechecking a call, make canCoerceFrom() allow any type to fit a type parameter ("$A"), and track the alias in a map. if the overall coercion works, use the map to "fill in the blanks" of the return type. also clean up disjoint type merging a bit, since we need to try to merge after filling in the blanks too.